% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdc_current_circulation.R
\name{algorand}
\alias{algorand}
\title{Fetch the current USDC circulation on Algorand}
\usage{
algorand()
}
\value{
the current circulating amount of USDC
}
\description{
Queries the `AlgoExplorer` explorer to retrieve the ASA balance of USDC
}
