% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdc_charts.R
\name{chart_current_supply_usdc}
\alias{chart_current_supply_usdc}
\title{Plot the current USDC circulation on all official Centre blockchains}
\usage{
chart_current_supply_usdc()
}
\value{
a ggplot2 chart showing the current circulating amount of USDC
}
\description{
Queries the Algorand, Ethereum, Solana, and Stellar blockchains to retrieve the token balances of USDC
on each blockchain and then plots the result on a bar chart
}
\examples{
chart_current_supply_usdc()
}
