% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdc_charts.R
\name{chart_historical_supply_usdc}
\alias{chart_historical_supply_usdc}
\title{Plot the historical USDC circulation on Ethereum}
\usage{
chart_historical_supply_usdc()
}
\value{
a ggplot2 chart showing the historical circulating amount of USDC on Ethereum
}
\description{
Queries the CoinMetrics API to retrieve the circulating supply of USDC
on Ethereum and then plots the result on a line chart
}
\examples{
chart_historical_supply_usdc()
}
