% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdc_current_circulation.R
\name{ethereum}
\alias{ethereum}
\title{Fetch the current USDC circulation on Ethereum}
\usage{
ethereum()
}
\value{
the current circulating amount of USDC
}
\description{
Queries the `blockchair` explorer to retrieve the erc-20 balance of USDC
}
