% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdc_historical.R
\name{ethereum_historical}
\alias{ethereum_historical}
\title{Fetch historical USDC metrics on Ethereum}
\usage{
ethereum_historical(metric = "CapMrktCurUSD")
}
\arguments{
\item{metric}{(character) A valid CoinMetrics asset metric}
}
\value{
a dataframe with the historical USDC metrics (see: https://docs.coinmetrics.io/api/v4#operation/getCatalogAllMetrics)
}
\description{
Queries the `CoinMetrics` API to retrieve historical USDC metrics
}
\examples{
ethereum_historical(metric = 'CapMrktCurUSD')
}
