% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdc-historical.R
\name{fetch_historical_ethereum}
\alias{fetch_historical_ethereum}
\title{Fetch historical USDC metrics on Ethereum}
\usage{
fetch_historical_ethereum(metric = "CapMrktCurUSD")
}
\arguments{
\item{metric}{(character) A valid CoinMetrics asset metric}
}
\value{
a dataframe with the historical USDC metrics (see: https://docs.coinmetrics.io/api/v4#operation/getCatalogAllMetrics)
}
\description{
Queries the `CoinMetrics` API to retrieve historical USDC metrics
}
\examples{
fetch_historical_ethereum(metric = "CapMrktCurUSD")
}
