% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdc-current.R
\name{fetch_supply_algorand}
\alias{fetch_supply_algorand}
\title{Fetch the current USDC circulation on Algorand}
\usage{
fetch_supply_algorand()
}
\value{
the current circulating amount of USDC
}
\description{
Queries the `AlgoExplorer` explorer to retrieve the ASA balance of USDC
}
\examples{
fetch_supply_algorand()
}
