% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdc_current.R
\name{fetch_supply_ethereum}
\alias{fetch_supply_ethereum}
\title{Fetch the current USDC circulation on Ethereum}
\usage{
fetch_supply_ethereum()
}
\value{
the current circulating amount of USDC
}
\description{
Queries the `blockchair` explorer to retrieve the erc-20 balance of USDC
}
\examples{
fetch_supply_ethereum()
}
