% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdc_current.R
\name{fetch_supply_solana}
\alias{fetch_supply_solana}
\title{Fetch the current USDC circulation on Solana}
\usage{
fetch_supply_solana()
}
\value{
the current circulating amount of USDC
}
\description{
Queries the `solana` JSON RPC api to retrieve the SPL token balance of USDC
}
\examples{
fetch_supply_solana()

}
