% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdc_current.R
\name{fetch_supply_stellar}
\alias{fetch_supply_stellar}
\title{Fetch the current USDC circulation on Stellar}
\usage{
fetch_supply_stellar()
}
\value{
the current circulating amount of USDC
}
\description{
Queries the `stellar` horizon explorer to retrieve the asset balance of USDC
}
\examples{
fetch_supply_stellar()
}
