% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdc-current.R
\name{fetch_supply_usdc}
\alias{fetch_supply_usdc}
\title{Fetch the current USDC circulation on all official Centre blockchains}
\usage{
fetch_supply_usdc()
}
\value{
a tibble containing the current circulating amount of USDC
}
\description{
Queries the Algorand, Ethereum, Solana, and Stellar blockchains to retrieve the token balances of USDC on each blockchain
}
\examples{
fetch_supply_usdc()
}
