% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-data.R
\docType{package}
\name{metrics}
\alias{metrics}
\title{Supported metrics available through Coin Metrics}
\description{
This dataset describes the field names that can be queried through
the Coin Metrics V4 API.  A description of each field provides context
for understanding the meaning of the metric.
}
