% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdc_current.R
\name{solana}
\alias{solana}
\title{Fetch the current USDC circulation on Solana}
\usage{
solana()
}
\value{
the current circulating amount of USDC
}
\description{
Queries the `solana` JSON RPC api to retrieve the SPL token balance of USDC
}
\examples{
solana()
}
