% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdc_current_circulation.R
\name{usdc_supply}
\alias{usdc_supply}
\title{Fetch the current USDC circulation on all official Centre blockchains}
\usage{
usdc_supply()
}
\value{
a tibble containing the current circulating amount of USDC
}
\description{
Queries the Algorand, Ethereum, Solana, and Stellar blockchains to retrieve the token balances of USDC on each blockchain
}
