% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdc_historical.R
\docType{data}
\name{coinmetrics_supported_metrics}
\alias{coinmetrics_supported_metrics}
\title{Supported CoinMetrics metrics for USDC}
\format{
An object of class \code{character} of length 26.
}
\usage{
coinmetrics_supported_metrics
}
\description{
Supported CoinMetrics metrics for USDC
}
\keyword{datasets}
